+++ src/main/java/at/hannibal2/skyhanni/features/gui/compacttablist/RenderColumn.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.features.misc.compacttablist
+package at.hannibal2.skyhanni.features.gui.compacttablist
 
 class RenderColumn(val lines: MutableList<TabLine> = mutableListOf()) {
