+++ src/main/java/at/hannibal2/skyhanni/features/gui/compacttablist/TabListRenderer.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.features.misc.compacttablist
+package at.hannibal2.skyhanni.features.gui.compacttablist
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.config.ConfigUpdaterMigrator
